refactor: S3 코드 통합, util 비우기, email 레이어 이름 통일 (#57)
#57 의 2·3단계.

S3 코드가 두 곳에 흩어져 있었다.
  s3/util/S3Uploader     업로드
  util/S3Downloader      ← 여기만 밖에 있었다
같은 일을 하는 코드가 다른 패키지에 있으면 '어디를 봐야 하나' 가 매번 생긴다.

util 패키지를 비웠다.
  util/CustomOAuth2SuccessHandler -> member/config/
util 은 '어디에 둘지 모르겠는 것' 의 다른 이름이 되기 쉽다.
인증 성공 핸들러는 member 도메인의 것이다.

email 만 레이어 이름이 달랐다(헥사고날 잔재).
  application  -> service
  presentation -> controller
나머지 7개 도메인과 맞춘다. 남겨두면 '어느 쪽이 우리 규칙인가' 가 흐려진다.

그리고 조사 문서에 채용 근거를 기록했다.
원티드 검색 API 직접 조회 결과 WebRTC 1건, LiveKit·OpenVidu·화상회의 0건.
제품 이름은 검색 키워드로서 가치가 없다.

테스트 212개 통과
