+++ docs/research/04-language-boundary.md
@@ -379,6 +379,30 @@ Discord로 실증됐고, **JVM은 정확히 그 지점에서 강하다.**
 
 ## 7. 국내 채용 실태 (2026-08 기준)
 
+### 화상/미디어 제품명으로는 뽑지 않는다 (2026-08-23 추가)
+
+원티드 검색 API 를 직접 조회했다.
+
+| 검색어 | 공고 수 |
+|---|---:|
+| **WebRTC** | **1** — 컨워스, *GPU 기반 실시간 스트리밍/렌더링 (Senior)* |
+| **LiveKit** | **0** |
+| **OpenVidu** | **0** |
+| **화상회의** | **0** (관련 회사는 있으나 공고 없음) |
+
+**제품 이름은 검색 키워드로서 가치가 없다.**
+WebRTC 조차 1건이고 그것도 **시니어 + GPU 렌더링**이라 신입 백엔드가 겨냥할 자리가 아니다.
+
+> 그래서 포트폴리오에서 *"OpenVidu 로 구현했습니다"* 는 아무 신호를 주지 못한다.
+> **판단을 말해야 한다** — *"1:N 방송에 SFU 를 쓰면 왜 안 되는지 대역폭으로 계산했다"*.
+> WebRTC 는 커넥션마다 SRTP 키가 달라 CDN 캐싱이 원천적으로 불가하고,
+> 그래서 CPU 보다 오리진 대역폭이 먼저 터진다. ([전송 비용 모델](../ops/02-egress-cost-model.md))
+
+**한계** — 원티드 한 곳, 한 시점이다. 사람인·잡코리아까지 보면 숫자가 달라질 수 있다.
+다만 **0건과 1건의 차이는 표본을 늘려도 잘 뒤집히지 않는다.**
+
+
+
 ### 7-1. SOOP (구 아프리카TV)
 
 | 직무 | 요구 언어 | 경력 |
