+++ src/main/java/com/edu/edumeet/email/dto/request/EmailRequest.java
@@ -1,4 +1,4 @@
-package com.edu.edumeet.email.presentation.dto.request;
+package com.edu.edumeet.email.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
