+++ src/main/java/com/edu/edumeet/s3/util/S3Downloader.java
@@ -1,4 +1,4 @@
-package com.edu.edumeet.util;
+package com.edu.edumeet.s3.util;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
